requirements.txt
alembic==1.14.0
imbalanced-learn==0.12.4
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.40.0
Deprecated==1.2.15
docker==7.1.0
Flask==3.1.0
fonttools==4.55.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
loguru==0.7.3
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mlflow==2.19.0
mlflow-skinny==2.19.0
numpy==1.26.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0

